API 요구사항에 맞게 전부 수정완료

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -2,6 +2,8 @@ package com.sprint.mission.discodeit.service.basic;
 
 import com.sprint.mission.discodeit.entity.*;
 import com.sprint.mission.discodeit.entity.BinaryContent;
+import com.sprint.mission.discodeit.entity.dto.binaryContentDto.BinaryContentRequestDto;
+import com.sprint.mission.discodeit.entity.dto.binaryContentDto.BinaryContentResponseDto;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.entity.dto.messageDto.*;
 import com.sprint.mission.discodeit.exception.InvalidInputException;
@@ -13,7 +15,9 @@ import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.*;
 import java.util.stream.Collectors;
 
@@ -26,51 +30,38 @@ public class BasicMessageService implements MessageService {
     private final MessageRepository messageRepository;
     private final BinaryContentRepository binaryContentRepository;
 
-    private List<UUID> saveAttachment(List<AttachmentDto> attachments) {
+    private List<UUID> saveAttachment(List<MultipartFile> attachments) {
         if (attachments == null || attachments.isEmpty()) {
             return new ArrayList<>();
         }
 
+        if (attachments.size() > 10) {
+            throw new InvalidInputException("파일은 한번에 10개까지만 보낼 수 있습니다."); // 예외 임시
+        }
+
         List<UUID> ids = new ArrayList<>();
 
-        for (AttachmentDto attachment : attachments) {
-            BinaryContent file = new BinaryContent(
-                    attachment.file(),
-                    attachment.fileName(),
-                    attachment.fileType()
-            );
-            binaryContentRepository.save(file);
+        for (MultipartFile attachment : attachments) {
+            try {
+                BinaryContent binaryContent = BinaryContent.builder()
+                        .binaryData(attachment.getBytes())
+                        .dataName(attachment.getOriginalFilename())
+                        .dataType(attachment.getContentType())
+                        .build();
 
-            ids.add(file.getId());
+                binaryContentRepository.save(binaryContent);
+                ids.add(binaryContent.getId());
+
+            } catch (IOException e) {
+                throw new RuntimeException("오류가 발생", e);
+            }
         }
         return ids;
     }
 
-    // Message Create
-    @Override
-    public MessageResponseDto createDirectMessage(DirectMessageRequestDto requestDto) {
+    public Message createMessage(MessageRequestDto requestDto, List<MultipartFile> attachments) {
         if ((requestDto.getContent() == null || requestDto.getContent().isBlank()) &&
-                (requestDto.getFiles() == null || requestDto.getFiles().isEmpty())) {
-            throw new InvalidInputException("공백을 보낼 수 없음");
-        }
-        User author = userRepository.findById(requestDto.getAuthorId())
-                .orElseThrow(() -> new NotFoundUserException("메시지를 보내는 사용자를 찾을 수 없음"));
-        User receiver = userRepository.findById(requestDto.getReceiverId())
-                .orElseThrow(() -> new NotFoundUserException("메시지를 받을 사용자를 찾을 수 없음"));
-
-        List<UUID> attachmentIds = saveAttachment(requestDto.getFiles());
-
-        Message message = new Message(author, receiver, requestDto.getContent(), attachmentIds);
-        messageRepository.save(message);
-
-        return MessageResponseDto.from(message);
-
-    }
-
-    @Override
-    public MessageResponseDto createChannelMessage(ChannelMessageRequestDto requestDto) {
-        if ((requestDto.getContent() == null || requestDto.getContent().isBlank()) &&
-                (requestDto.getFiles() == null || requestDto.getFiles().isEmpty())) {
+                (attachments == null || attachments.isEmpty())) {
             throw new InvalidInputException("공백을 보낼 수 없음");
         }
 
@@ -79,57 +70,40 @@ public class BasicMessageService implements MessageService {
         Channel channel = channelRepository.findById(requestDto.getChannelId())
                 .orElseThrow(() -> new NotFoundChannelException("메시지를 받을 채널을 찾을 수 없음"));
 
-        // 텍스트만 보내면 getFiles == null
-        if (requestDto.getFiles() != null && requestDto.getFiles().size() > 10) {
-            throw new InvalidInputException("파일은 한번에 10개까지만 보낼 수 있습니다."); // 예외 임시
-        }
+        List<UUID> attachmentIds = saveAttachment(attachments);
 
-        List<UUID> attachmentIds = saveAttachment(requestDto.getFiles());
-        Message message = new Message(author, channel, requestDto.getContent(), attachmentIds);
+        Message message = Message.builder()
+                .authorId(requestDto.getAuthorId())
+                .channelId(requestDto.getChannelId())
+                .content(requestDto.getContent())
+                .attachmentIds(attachmentIds)
+                .build();
         messageRepository.save(message);
 
-        return MessageResponseDto.from(message);
-    }
-
-
-    // Message Read
-    @Override
-    public MessageResponseDto findMessageById(UUID messageId) {
-        Message message = messageRepository.findById(messageId)
-                .orElseThrow(() -> new NoSuchElementException("해당 메시지를 찾을 수 없습니다."));
-
-        return MessageResponseDto.from(message);
-    }
-
-    // update를 해도 순서는 바뀌지않음 생성일자로 정렬
-    @Override
-    public List<MessageResponseDto> findMessageBetweenUsers(UUID userId1, UUID userId2) {
-        return messageRepository.findAllByBetweenUserIds(userId1, userId2)
-                .stream().sorted(Comparator.comparing(Message::getCreatedAt))
-                .map(MessageResponseDto::from).collect(Collectors.toList());
+        return message;
     }
 
     @Override
-    public List<MessageResponseDto> findAllByChannelId(UUID channelId) {
+    public List<Message> findAllByChannelId(UUID channelId) {
         return messageRepository.findAllByChannelId(channelId).stream()
                 .sorted(Comparator.comparing(Message::getCreatedAt))
-                .map(MessageResponseDto::from).collect(Collectors.toList());
+                .collect(Collectors.toList());
     }
 
     // Message Update
     @Override
-    public Optional<MessageResponseDto> updateMessage(MessageUpdateDto updateDto) {
-        if (updateDto.newContent() == null || updateDto.newContent().isBlank()) {
-            deleteMessage(updateDto.messageId());
-            return Optional.empty();
-        }
+    public Message updateMessage(UUID messageId, MessageUpdateDto updateDto) {
 
-        Message message = messageRepository.findById(updateDto.messageId())
+        Message message = messageRepository.findById(messageId)
                 .orElseThrow(() -> new NoSuchElementException("메시지를 찾을 수 없습니다."));
+
+        if (updateDto.newContent() == null || updateDto.newContent().isBlank()) {
+            throw new InvalidInputException("비어 있을 수 없습니다.");
+        }
 
         message.updateContent(updateDto.newContent());
         messageRepository.save(message);
-        return Optional.of(MessageResponseDto.from(message));
+        return message;
     }
 
     // Message Delete
@@ -138,16 +112,12 @@ public class BasicMessageService implements MessageService {
         Message message = messageRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("메시지를 찾을 수 없습니다."));
 
-        for (UUID attachmentId : message.getAttachmentIds()){
-            binaryContentRepository.deleteById(attachmentId);
+        if (message.getAttachmentIds() != null && !message.getAttachmentIds().isEmpty()) {
+            for (UUID attachmentId : message.getAttachmentIds()) {
+                binaryContentRepository.deleteById(attachmentId);
+            }
         }
 
         messageRepository.deleteById(id);
-    }
-
-    @Override
-    public List<MessageResponseDto> findAll() {
-        return messageRepository.findAll().stream()
-                .map(MessageResponseDto::from).collect(Collectors.toList());
     }
 }
